Split generator exercises into 3 modules instead of 2
